diplom/requirements.txt
amqp==5.1.1
anyio==4.0.0
APScheduler==3.6.3
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.1.0
binaryornot==0.4.4
cachetools==4.2.2
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.4.0
cryptography==41.0.5
daphne==4.0.0
DateTime==5.2
Django==4.2.7
django-admin-interface==0.26.1
django-colorfield==0.10.1
django-cors-headers==4.3.0
django-environ==0.11.2
django-grappelli==3.0.8
django-redis==5.4.0
django-tailwind==3.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
Jinja2==3.1.2
kombu==5.3.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-memcached==1.59
python-slugify==8.0.1
python-telegram-bot==13.7
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
reportlab==4.0.6
requests==2.31.0
rich==13.6.0
routing==0.2.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
telegram==0.0.1
text-unidecode==1.3
tornado==6.3.3
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.9
whitenoise==6.6.0
zope.interface==6.1
